ai_service/requirements.txt
fastapi
uvicorn
numpy
pandas
scikit-learn
pydantic
requests
ultralytics
Pillow
python-multipart
langchain
langchain-community
langchain-openai
faiss-cpu
pypdf
tiktoken
python-dotenv
